Apply changes based on feedback
- Rename from `Methods/NullableWhitespace` to `Functions/NullableTypeDeclaration` to be more descriptive
- Consider more situations then only a single whitespace - for non-expected non-whitespace tokens I opted not to fix
- Deduplicated test situations
- Added entries to `ruleset.xml` and `packages.xml`
