+++ src/Standards/PSR12/Docs/Functions/NullableTypeDeclarationStandard.xml
@@ -1,4 +1,4 @@
-<documentation title="Nullable Whitespace Keywords">
+<documentation title="Nullable Type Declarations Functions">
     <standard>
         <![CDATA[
     In nullable type declarations there MUST NOT be a space between the question mark and the type.
@@ -13,8 +13,17 @@ public function functionName(?string $arg1, ?int $arg2): ?string
         ]]>
         </code>
         <code title="Invalid: superfluous whitespace used.">
-            <![CDATA[
+        <![CDATA[
 public function functionName(? string $arg1, ? int $arg2): ? string
+{
+}
+        ]]>
+        </code>
+        <code title="Invalid: unexpected characters used.">
+            <![CDATA[
+public function functionName(? /* nullable for a reason */ string $arg1): ?
+    // nullable for a reason
+    string
 {
 }
         ]]>
