--- src/Standards/PSR12/Sniffs/Methods/NullableWhitespaceSniff.php
@@ -1,57 +1,0 @@
-<?php
-/**
- * Verifies that nullable typehints are lacking superfluous whitespace, e.g. ?int
- *
- * @author    Greg Sherwood <gsherwood@squiz.net>
- * @copyright 2006-2015 Squiz Pty Ltd (ABN 77 084 670 600)
- * @license   https://github.com/squizlabs/PHP_CodeSniffer/blob/master/licence.txt BSD Licence
- */
-
-namespace PHP_CodeSniffer\Standards\PSR12\Sniffs\Methods;
-
-use PHP_CodeSniffer\Files\File;
-use PHP_CodeSniffer\Sniffs\Sniff;
-
-class NullableWhitespaceSniff implements Sniff
-{
-
-
-    /**
-     * Returns an array of tokens this test wants to listen for.
-     *
-     * @return array
-     */
-    public function register()
-    {
-        return [T_NULLABLE];
-
-    }//end register()
-
-
-    /**
-     * Processes this test, when one of its tokens is encountered.
-     *
-     * @param \PHP_CodeSniffer\Files\File $phpcsFile The file being scanned.
-     * @param int                         $stackPtr  The position of the current token in the
-     *                                               stack passed in $tokens.
-     *
-     * @return void
-     */
-    public function process(File $phpcsFile, $stackPtr)
-    {
-        $tokens = $phpcsFile->getTokens();
-        if (isset($tokens[($stackPtr + 1)]) === true
-            && $tokens[($stackPtr + 1)]['code'] === T_WHITESPACE
-            && $phpcsFile->addFixableError(
-                'Superfluous whitespace after nullable',
-                ($stackPtr + 1),
-                'WhitespaceFound'
-            ) === true
-        ) {
-            $phpcsFile->fixer->replaceToken(($stackPtr + 1), '');
-        }
-
-    }//end process()
-
-
-}//end class
