--- src/Standards/PSR12/Tests/Methods/NullableWhitespaceUnitTest.inc
@@ -1,41 +1,0 @@
-<?php
-
-class MyClass {
-    public function validReturnType(): ?MyObject
-    {
-    }
-
-    public function validReturnTypeForThisSniff():?MyObject
-    {
-    }
-
-    public function invalidReturnTypeTooLittleWhitespace():? MyObject
-    {
-    }
-
-    public function invalidReturnTypeWhitespaceAfterQuestionMark(): ? MyObject
-    {
-    }
-
-    public function validParamType(?MyObject $object)
-    {
-    }
-
-    public function invalidParamTypeTooMuchWhitespaceAfterQuestionMark(? MyObject $object)
-    {
-    }
-}
-
-interface MyInterface {
-    public function validReturnType(): ?MyObject;
-
-    public function validReturnTypeForThisSniff():?MyObject;
-
-    public function invalidReturnTypeTooLittleWhitespace():? MyObject;
-
-    public function invalidReturnTypeWhitespaceAfterQuestionMark(): ? MyObject;
-
-    public function validParamType(?MyObject $object);
-
-    public function invalidParamTypeTooMuchWhitespaceAfterQuestionMark(? MyObject $object);
-}
